+++ frontend/src/App.js
@@ -5,7 +5,7 @@ import LoginChoice from "./pages/LoginChoice";
 import Layout from "./pages/layout/Layout";
 import Setting from "./pages/Setting";
 import Main from "./pages/Main";
-import History from "./pages/History";
+import History from "./components/history/main/page/History.jsx";
 import InventoryView from "./pages/InventoryView";
 import Income from "./pages/Income";
 import SignUp from "./pages/SignUp.jsx";
@@ -14,23 +14,24 @@ import SignUpBroker from "./components/login_signup/pagecomponent/SignUpBroker.j
 import SignUpBrokerChoice from "./components/login_signup/pagecomponent/SignUpBrokerChoice.jsx";
 import SignUpInfo from "./components/login_signup/pagecomponent/SignUpInfo.jsx";
 import MainSetting from "./components/setting/main/MainSetting";
-import OutgoingList from "./pages/OutgoingList";
+import OutgoingList from "./components/history/outgoing/page/OutgoingList.jsx";
 import QrInfo from "./components/setting/qrinfo/QrInfo";
 import UserInfoSetting from "./components/setting/userinfo/UserInfoSetting";
 import PasswordChange from "./components/setting/userinfo/password_change/PasswordChange";
 import TeamInfo from "./components/setting/teaminfo/TeamInfo";
 import InventorySetting from "./components/setting/inventoryinfo/InventorySetting";
 
-import OutgoingReqList from "./pages/OutgoingReqList";
+import OutgoingReqList from "./components/history/outgoing/page/OutgoingReqList.jsx";
 import Alarm from "./components/setting/alarm/Alarm";
-import IncomingList from "./pages/IncomingList";
-import AdjustList from "./pages/AdjustList";
+import IncomingList from "./components/history/incoming/page/IncomingList.jsx";
+import AdjustList from "./components/history/adjust/page/AdjustList.jsx";
 import AlarmHistory from "./components/common/header/AlarmHistory";
 import Cart from "./pages/Cart";
 import BrokerMain from "./components/main/broker/BrokerMain";
 import TradeMain from "./components/trade/TradeMain";
 import TradeView from "./pages/TradeView";
 import TradeBrokerMain from "./components/trade/TradeBrokerMain";
+import HistoryOutlet from "./pages/HistoryOutlet.jsx";
 
 function App() {
 	return (
@@ -60,8 +61,24 @@ function App() {
 						<Route path="/main" element={<Main />}>
 							<Route path="" element={<BrokerMain />} />
 						</Route>
-						<Route path="/history" element={<History />} />
-						{/* histories */}
+
+						<Route path="/histories" element={<HistoryOutlet />}>
+						<Route path="" element={<History />} />
+						<Route
+							path="outgoings"
+							element={<OutgoingList />}
+						/>
+						<Route
+							path="incomingList"
+							element={<IncomingList />}
+						/>
+						<Route
+							path="outgoings/requests"
+							element={<OutgoingReqList />}
+						/>
+						<Route path="adjusts" element={<AdjustList />} />
+					</Route>
+						
 						<Route
 							path="/inventories"
 							element={<InventoryView />}
@@ -76,41 +93,26 @@ function App() {
 						</Route>
 						<Route path="/cart" element={<Cart />} /> {/* carts */}
 						<Route path="/incoming" element={<Income />} />
-						<Route
-							path="/outgoingList"
-							element={<OutgoingList />}
-						/>
-						{/* /outgoings */}
-						<Route
-							path="/outgoingReqList"
-							element={<OutgoingReqList />}
-						/>
-						{/* /outgoings/requests */}
-						<Route
-							path="/incomingList"
-							element={<IncomingList />}
-						/>
-						<Route path="/adjustList" element={<AdjustList />} />
-						{/* /adjusts */}
+						
+						
+						
 						<Route
 							path="/alarm-history"
 							element={<AlarmHistory />}
 						/>
 					</Route>
-					<Route path="/login" element={<LoginChoice />} />{" "}
-					{/* /login */}
+					<Route path="/login" element={<LoginChoice />} />
 					<Route path="/signup" element={<SignUp />}>
 						<Route path="" element={<SignUpInfo />} />
 						<Route path="broker" element={<SignUpBroker />} />
-						<Route path="choice" element={<SignUpChoice />} />{" "}
-						{/* /signup/choice */}
+						<Route path="choice" element={<SignUpChoice />} />
 						<Route
 							path="broker-choice"
 							element={<SignUpBrokerChoice />}
-						/>{" "}
-						{/* /signup/broker-choice */}
-					</Route>{" "}
-					{/* /signup */}
+						/>
+					</Route>
+					
+					
 				</Routes>
 			</BrowserRouter>
 		</div>
