+++ frontend/src/components/history/main/itemcomponent/Period.jsx
