+++ frontend/src/components/history/main/itemcomponent/PeriodEnd.jsx
