+++ frontend/src/components/history/outgoing/itemcomponent/OutgoingDelete.jsx
