+++ frontend/src/components/history/outgoing/itemcomponent/OutgoingItemDetails.jsx
