+++ frontend/src/components/history/outgoing/itemcomponent/OutgoingListHeader.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ReactComponent as Sort } from '../../assets/history/sort-icon.svg'; 
+
 
 const OutgoingListHeader = () => {
     return (
