--- frontend/src/components/incoming/styles/IncomingListItem.css
@@ -1,38 +1,0 @@
-/* styles/IncomingListItem.css */
-
-/* 텍스트를 감싸는 컨테이너의 위치 조정 */
-.relative {
-    display: flex;
-    align-items: flex-start; /* 아이템들을 상단으로 정렬 */
-    padding-top: 10px; /* 상단 패딩을 추가하여 텍스트를 위로 이동 */
-}
-
-/* 텍스트 컨테이너의 위치 조정 */
-.text-container {
-    position: relative; /* 위치를 상대적으로 설정 */
-    margin-top: 10px; /* 상단 여백을 추가하여 텍스트를 올림 */
-}
-
-/* 텍스트의 위치를 조정 */
-.truncate-text {
-    display: inline-block;
-    max-width: 10em; /* 한글 10글자 정도의 너비로 설정 */
-    white-space: nowrap; /* 텍스트가 한 줄로 유지되도록 설정 */
-    overflow: hidden;
-    text-overflow: ellipsis;
-    vertical-align: middle; /* 텍스트를 수직 가운데 정렬 */
-}
-
-/* 날짜 텍스트 스타일 */
-.date-text {
-    position: absolute;
-    top: 5px; /* 이미지 위쪽으로 위치를 조정 */
-    left: 10px; /* 왼쪽으로 정렬 */
-    text-align: left; /* 텍스트를 왼쪽 정렬 */
-    background-color: rgba(255, 255, 255, 0.75); /* 배경을 반투명하게 설정 */
-    padding: 2px 5px;
-    font-size: 14px;
-    border-radius: 5px;
-    font-weight: bold; /* 날짜를 강조하기 위해 볼드체로 설정 */
-    white-space: nowrap; /* 텍스트가 줄 바꿈 없이 한 줄로 유지되도록 설정 */
-}
