+++ frontend/src/pages/HistoryOutlet.jsx
@@ -1,0 +1,11 @@
+import { Outlet } from "react-router-dom";
+
+const HistoryOutlet = () => {
+	return (
+		<div className="flex flex-col justify-center">
+			<Outlet />
+		</div>
+	);
+};
+
+export default HistoryOutlet;
